Code cleanup.

+++ src/main/java/shadows/plants2/block/forgotten/BlockCrystalTorch.java
@@ -49,9 +49,9 @@ public class BlockCrystalTorch extends BlockTorch implements IHasModel, IHasReci
 	@Override
 	public void randomDisplayTick(IBlockState state, World world, BlockPos pos, Random rand) {
 		EnumFacing enumfacing = state.getValue(FACING);
-		double d0 = (double) pos.getX() + 0.5D;
-		double d1 = (double) pos.getY() + 0.7D;
-		double d2 = (double) pos.getZ() + 0.5D;
+		double d0 = pos.getX() + 0.5D;
+		double d1 = pos.getY() + 0.7D;
+		double d2 = pos.getZ() + 0.5D;
 
 		ParticleWhiteFlame p;
